add a review shell
there is so much spam in lighthouse - I'm not going to permit the import
process to propogate it. Make it possible to review all tickets, and
then review all comments. It's possible to whitelist/mark-as-spammer a
user so that it's only necessary to review each user once (if the user's
activity makes it clear that they belong to one or the other group).